requirements.txt
fastapi>=0.104.1,<0.110.0
uvicorn>=0.24.0,<0.30.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
sqlalchemy>=2.0.23,<2.1.0
alembic>=1.13.1,<2.0.0
python-multipart>=0.0.6,<0.1.0
jinja2>=3.1.2,<4.0.0
openai>=1.3.7,<2.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.2,<5.0.0
python-dotenv>=1.0.0,<2.0.0
streamlit>=1.28.1,<2.0.0
pandas>=1.5.0,<3.0.0
markdown>=3.5.1,<4.0.0
python-docx>=1.1.0,<2.0.0
python-pptx>=0.6.23,<1.0.0
PyJWT>=2.8.0,<3.0.0
